requirements.txt
requests
python-dotenv
pytz
selenium
pyotp